+++ src/person/person.controller.ts
@@ -4,6 +4,7 @@ import Person from './person.model.js'
 import { MongoServerError } from 'mongodb'
 import { Result, validationResult } from 'express-validator'
 import { PaginateOptions } from 'mongoose'
+import Competitor from '../competitor/competitor.model.js'
 
 export async function findAll(req: Request, res: Response) {
   const result: Result = validationResult(req)
@@ -157,6 +158,12 @@ export async function remove(req: Request, res: Response) {
   }
 
   try {
+    const competitors = await Competitor.find({ person: id })
+
+    if (competitors.length !== 0) {
+      return res.status(409).send({ message: 'El socio tiene inscripciones cargadas' })
+    }
+
     const person = await Person.findByIdAndDelete(id)
 
     if (!person) {
